PUT, DELETE 요청 시 컨트롤러에서 해당 트윗을 작성한 유저가 맞는 지 확인해서 맞으면 요청을 수행하고 그렇지 않으면 400번 대 에러 응답하도록 수정

+++ server/controller/tweet.js
@@ -34,6 +34,11 @@ export const create = async (req, res) => {
 export const update = async (req, res) => {
   const tweetId = req.params.id;
   const textData = req.body.text;
+
+  const getTweet = await tweetRepository.getId(tweetId);
+  if (!getTweet) return res.sendStatus(404);
+  if (getTweet.userId !== req.body.userId) return res.sendStatus(403);
+
   const findTweet = await tweetRepository.update(tweetId, textData);
 
   if (findTweet) {
@@ -45,6 +50,11 @@ export const update = async (req, res) => {
 
 export const remove = async (req, res) => {
   const tweetId = req.params.id;
+
+  const getTweet = await tweetRepository.getId(tweetId);
+  if (!getTweet) return res.sendStatus(404);
+  if (getTweet.userId !== req.body.userId) return res.sendStatus(403);
+
   await tweetRepository.remove(tweetId);
   res.sendStatus(204); // 삭제는 데이터를 삭제하는 것이기 때문에 삭제가 성공 했는 지 알맞은 상태 코드만 보낸다.
 };
